# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  realNumberDNOClone_internal_models.BulkQueryRequest:
    properties:
      channel:
        type: string
      phoneNumbers:
        items:
          type: string
        type: array
    type: object
  realNumberDNOClone_internal_models.BulkQueryResponse:
    properties:
      hits:
        type: integer
      misses:
        type: integer
      results:
        items:
          $ref: '#/definitions/realNumberDNOClone_internal_models.DNOQueryResponse'
        type: array
      total:
        type: integer
    type: object
...
  realNumberDNOClone_internal_models.DNOQueryResponse:
    properties:
      channel:
        type: string
      dataset:
        type: string
      isDno:
        type: boolean
      lastUpdated:
        type: string
      phoneNumber:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: FakeNumber DNO
    url: https://github.com/drewkosta/realNumberDNOClone
  description: |-
    Do Not Originate database API for preventing illegal robocalls and caller ID spoofing.
    Carriers query the DNO database in real-time to block spoofed calls.
  title: FakeNumber DNO API
  version: "1.0"
paths:
...
  /dno/query:
    get:
      consumes:
      - application/json
      description: Check if a phone number is on the Do Not Originate list
      parameters:
      - description: 10-digit phone number
        example: "5551234567"
        in: query
        name: phoneNumber
        required: true
        type: string
      - default: voice
        description: voice or text
        in: query
        name: channel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.DNOQueryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - APIKeyAuth: []
        BearerAuth: []
      summary: Single DNO lookup
      tags:
      - Query
  /dno/query/bulk:
    post:
      consumes:
      - application/json
      description: Check up to 1000 phone numbers against the DNO list in a single
        request
      parameters:
      - description: Phone numbers and channel
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/realNumberDNOClone_internal_models.BulkQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.BulkQueryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - APIKeyAuth: []
        BearerAuth: []
      summary: Bulk DNO lookup
      tags:
      - Query
...
securityDefinitions:
  APIKeyAuth:
    description: Organization API key for machine-to-machine access
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT Bearer token (prefix with "Bearer ")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"